+++ Frontend/src/components/organisms/overlays/wizard/steps/StepConfirmation.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-const StepBudgetInfo: React.FC = () => {
+const StepBudgetIncome: React.FC = () => {
   return (
     <div className="text-center">
       <h3 className="text-xl font-semibold mb-2">Budget Info (Placeholder)</h3>
@@ -9,4 +9,4 @@ const StepBudgetInfo: React.FC = () => {
   );
 };
 
-export default StepBudgetInfo;
+export default StepBudgetIncome;
